moved everything to util/log.py and made output logger a child of the dbgym logger
